% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_freq.R
\docType{data}
\name{allele.frequencies.no.extra}
\alias{allele.frequencies.no.extra}
\title{Allele frequency Data.}
\format{The first column is the name of 678 different loci, the third column includes the allele frequencies of four
populations(Europe, America, Africa, Latino) of different alleles at this locus.}
\usage{
allele.frequencies.no.extra
}
\description{
A dataset containing the allele frequencies of Europe, America, Africa, Latino populations at 678 different loci.
}
\references{
Boca, S. M., & Rosenberg, N. A. (2011). Mathematical properties of Fst between admixed populations 
and their parental source populations. Theoretical population biology, 80(3), 208-216.
}
\keyword{datasets}
