% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_freq.R
\docType{data}
\name{allele.frequency}
\alias{allele.frequency}
\title{Allele frequency Data created by us.}
\format{The first column is the name of 678 different loci, the third column includes the allele frequencies of four
populations(Europe, America, Africa, Latino) of different alleles at this locus.}
\usage{
allele.frequency
}
\description{
A dataset containing the allele frequencies of Europe, America, Africa, Latino populations at 678 different loci. This 
dataset is the same as the one from the paper.
}
\references{
Wang, S., Ray, N., Rojas, W., Parra, M. V., Bedoya, G., Gallo, C., ... & Camrena, B. (2008). 
Geographic patterns of genome admixture in Latin American Mestizos. PLoS genetics, 4(3), e1000037.
}
\keyword{datasets}
