% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{closeStreams}
\alias{closeStreams}
\title{Explicitly close SciDB streams}
\usage{
closeStreams()
}
\value{
\code{NULL}
}
\description{
You should almost never need to call this in practice. The \code{\link{map}}
function automatically closes SciDB streams when done.
}

