% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aashind.R
\name{component_aashind}
\alias{component_aashind}
\title{AASHTO Group Index for NASIS/SSURGO components}
\usage{
component_aashind(components, FUN, ...)
}
\arguments{
\item{components}{A \code{SoilProfileCollection} Object containing low, 
RV, and high values of the standard inputs and SSURGO/NASIS-naming
convention to \code{ptf_aashind}. E.g. \code{sieveno200_l}, \code{ll_r},
\code{pi_h}, \code{aashtocl}, etc.}

\item{FUN}{A function to apply to the end result. Usually \code{floor}.}

\item{...}{Additional arguments to \code{FUN}.}
}
\value{
A \code{data.frame} containing L, RV and H calculated AASHTO Group Index ('calc_aashind_l','calc_aashind_r', 'calc_aashind_h').
}
\description{
Runs AASHTO Group Index (\code{ptf_aashind}) calculations on all horizons in a set of NASIS/SSURGO components (SoilProfileCollection).
}
\examples{
component_aashind(my.component.spc, floor).
}
\seealso{
\code{\link{SDA_query}}, \code{\link{fetchNASIS_components}}
}
\author{
Andrew G. Brown.
}
