% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_S1.R
\docType{data}
\name{dataset_S1}
\alias{dataset_S1}
\title{Raw dataset from the allele frequency dataset}
\format{The first line should be the name of the markers, and the first column contains the names of the populations, 
other columns should contain the genotypes(measured in base pairs).}
\usage{
dataset_S1
}
\description{
A dataset containing dataset should contain the genotypes for the markers.
}
\references{
Wang, S., Ray, N., Rojas, W., Parra, M. V., Bedoya, G., Gallo, C., ... & Camrena, B. (2008). 
Geographic patterns of genome admixture in Latin American Mestizos. PLoS genetics, 4(3), e1000037.
}
\keyword{datasets}
