% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainST.R
\name{explainST}
\alias{explainST}
\title{Explain a subgroup level taxa from Soil Taxaonomy}
\usage{
explainST(x)
}
\arguments{
\item{x}{a single subgroup, matching is exact and case insensitive}
}
\value{
a block of text, suitable for display in fixed-width font
}
\description{
Explain a subgroup level taxa from Soil Taxaonomy
}
\note{
This function currently accepts only subgroup taxa. There are plans to extend to arbitrary levels of the heirarchy.
}
