% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fst-function.R
\name{get.Fst.admix.general}
\alias{get.Fst.admix.general}
\title{Calculate Fst for pop.1 and admixed population.}
\usage{
get.Fst.admix.general(pops, gamma)
}
\arguments{
\item{pops}{Matirx.}

\item{gamma}{Numeric vectors.}
}
\value{
If all inputs are positive and logical, then the output
  will be an numeric vector.If there are missing values in the
  inputs, the output will be NA with a warning.
}
\description{
\code{get.Fst.admix.general} returns Fst for pop.1 and admixed population for the general case of K founding populations.
}
\details{
The input pops is matrix with K rows and I columns (sums on rows is 1), and
gamma are the ancestry fractions corresponding to the K founding populations
(add up to 1). Ht is the expected heterozygosity of the overall population,
and Hs is the mean expected heterozygosity across subpopulations. If the sums
on rows in the matrix is not 1 or the sum of vectors of gamma is not 1, then
cannot use this function
}
\examples{
pops <- matrix(nrow = 2, ncol = 10, c(0.059, 0.153, 0.132, 0.101, 0.08, 0.055, 0.106, 0.145,
0.048, 0.121, 1.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000), byrow=TRUE)
gamma <- c(0.3, 0.7)
get.Fst.admix.general(pops, gamma)
}
\references{
Boca, S. M., & Rosenberg, N. A. (2011). Mathematical properties of Fst between admixed populations 
and their parental source populations. Theoretical population biology, 80(3), 208-216.
}
\keyword{Fst}
