% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fst-function.R
\name{get.Fst.admix}
\alias{get.Fst.admix}
\title{Calculate Fst: between a founder population and the admixed population.}
\usage{
get.Fst.admix(pop.1, pop.2, gamma)
}
\arguments{
\item{pop.1}{Numeric vector.}

\item{pop.2}{Numeric vector.}

\item{gamma}{Numeric vector.}
}
\value{
If all inputs are positive and logical, then the output
  will be an numeric vector. If there are missing values in the
  inputs, the output will be NA with a warning.
}
\description{
\code{get.Fst.admix} returns the Fst between a founder population and the admixed population for the K=2 founder population case.
}
\details{
The input pop.1 and pop.2 represent allele frequencies in the same locus for populations 1
and 2 respectively, and gamma is the ancestry fraction corresponding to population 1. Ht is
the expected heterozygosity of the overall population, and Hs is the mean expected
heterozygosity across subpopulations.
}
\examples{
pop.1 <- 0.3
pop.2 <- 0.4
gamma <- 0.25
get.Fst.admix(pop.1, pop.2, gamma)
}
\references{
Boca, S. M., & Rosenberg, N. A. (2011). Mathematical properties of Fst between admixed populations 
and their parental source populations. Theoretical population biology, 80(3), 208-216.
}
\keyword{Fst}
