% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosity-function.R
\name{get.Hadm.admix.general}
\alias{get.Hadm.admix.general}
\title{Calculate the heterozygosity of the admixed population}
\usage{
get.Hadm.admix.general(pops, gamma)
}
\arguments{
\item{pops}{Matrix}

\item{gamma}{Numeric vectors}
}
\value{
If there are missing values in the inputs, the output will be NA with a warning. If the sums of rows in martix is not
  1 or the sum of gamma is not 1, it will stop.
}
\description{
\code{get.Hadm.admix.general} returns the heterozygosity of the admixed population.
}
\details{
The input is a matrix with K rows and I columns (sums on rows is 1), vector of gammas (add up to 1).
}
\examples{
pops <- matrix(nrow = 2, ncol = 10, c(0.059, 0.153, 0.132, 0.101, 0.08, 0.055, 0.106, 0.145, 0.048, 
               0.121, 1.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000, 0.000), byrow=TRUE)
gamma <- c(0.3, 0.7)
get.Hadm.admix.general(pops, gamma)
}
