% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygosity-function.R
\name{get.het}
\alias{get.het}
\title{Calculate the heterozygosity.}
\usage{
get.het(allele.freq)
}
\arguments{
\item{allele.freq}{Numeric vectors.}
}
\value{
If all inputs are positive and logical, then the output
  will be an numeric vector. If there are missing values in the
  inputs, the output will be NA with a warning.
}
\description{
\code{get.het} returns the heterozygosity of a population at a locus.
}
\details{
The input is the vector of allele ferquencies (should add up to 1).
  If there are missing values in the inputs or the sum of the inputs don't
  equal to 1, then cannot use this function.
}
\examples{
c <- c(0.2, 0.3, 0.5)
get.het(c)
}
\references{
Boca, S. M., & Rosenberg, N. A. (2011). Mathematical properties of Fst between admixed populations 
and their parental source populations. Theoretical population biology, 80(3), 208-216.
}
\keyword{heterozygosity}
