% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndCacheCalculation.R
\name{getAndCacheCalculation}
\alias{getAndCacheCalculation}
\title{getAndCacheCalculation}
\usage{
getAndCacheCalculation(calcName, fileName = paste0(calcName, ".txt"),
  basePath)
}
\arguments{
\item{calcName}{The name of the calculation in NASIS.}

\item{fileName}{The name of the file to save the calculation in. default: \code{calcName}.txt.}

\item{basePath}{The path to a directory where calculation file will be stored.}
}
\value{
Calculation CVIR code is written to specified file.
}
\description{
Query NASIS database by calculation name to store a flat-file (.txt) containing the CVIR code.
}
\examples{
getAndCacheCalculation(
   calcName = '15 AASHTO Group Index', 
   basePath = 'NASIS_calculations/'
)
}
\author{
D.E. Beaudette
}
