% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{getChunk}
\alias{getChunk}
\title{Obtain a single chunk from SciDB, returning \code{output}.}
\usage{
getChunk(output = list())
}
\arguments{
\item{output}{A list with chunk output to return to SciDB.}
}
\value{
an R list of values representing one SciDB chunk
}
\description{
A low-level, chunk-by-chunk interface to the SciDB streaming API.
Use this with the optional extra array argument in the SciDB stream
operator to read data before running \code{map}.
}
\seealso{
\code{\link{map}} \code{\link{closeStream}}
}

