% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formative-elements.R
\name{GreatGroupFormativeElements}
\alias{GreatGroupFormativeElements}
\title{Identify greatgroup formative elements.}
\usage{
GreatGroupFormativeElements(x)
}
\arguments{
\item{x}{subgroup or lower taxa}
}
\description{
Identify greatgroup formative elements.
}
