\name{gts.listRegion}
\alias{gts.listRegion}
\title{
Function querying all the geological time concept of a region or international geological time concept.
}
\description{Function querying all the geological time concept of a region or international geological time concept.
}
\usage{
gts.listRegion(prefix = NULL, graph = NULL)
}
\arguments{
\item{prefix}{
[character] prefix for SPARQL querying. [Optional]
}
\item{graph}{
[character] GRAPH for SPARQL querying. [Optional]
}
}

\references{
%% ~put references to the literature/web site here ~
}

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
gts.listRegion()
}