% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formative-elements.R
\name{isValidST}
\alias{isValidST}
\title{Check for valid order | suborder | greatgroup | subgroup.}
\usage{
isValidST(needle, level)
}
\arguments{
\item{needle}{vector of taxa}

\item{level}{single level of Soil Taxonomy hierarchy: order, suborder, greatgroup, subgroup}
}
\value{
logical vector, same lenght as needle
}
\description{
This function can be used to quickly test for valid levels within the 
Soil Taxonomy hierarchy, a defined within the 13th edition of The Keys to Soil Taxonomy.
Matches are case insensitive.
}
