% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchParentTaxa.R
\name{matchParentTaxa}
\alias{matchParentTaxa}
\title{Match Parent Taxon / Taxa}
\usage{
matchParentTaxa(needle, haystack, qgram.size = 4)
}
\arguments{
\item{needle}{search term, usually something like 'xeralfs'}

\item{haystack}{word list}

\item{qgram.size}{size of the q-gram, passed to \code{stringdist::stringdist()}}
}
\value{
first match
}
\description{
Match the parent taxa (multiple) or taxon (single) via string distance computed from formative element lists.
}
