% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_freq.R
\docType{data}
\name{nr.observed.alleles.no.extra}
\alias{nr.observed.alleles.no.extra}
\title{Observed count of alleles Data.}
\format{The first column is the name of 678 different loci, the third column includes the observed counts of different
alleles of four populations(Europe, America, Africa, Latino) at this locus.}
\usage{
nr.observed.alleles.no.extra
}
\description{
A dataset containing the observed counts of Europe, America, Africa, Latino populations at 678 different loci.
}
\references{
Boca, S. M., & Rosenberg, N. A. (2011). Mathematical properties of Fst between admixed populations 
and their parental source populations. Theoretical population biology, 80(3), 208-216.
}
\keyword{datasets}
