% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formative-elements.R
\name{OrderFormativeElements}
\alias{OrderFormativeElements}
\title{Extract Soil Order Formative Elements}
\usage{
OrderFormativeElements(x)
}
\arguments{
\item{x}{vector of soil order taxa}
}
\value{
a \code{list}
}
\description{
Extract Soil Order Formative Elements
}
\author{
D.E. Beaudette
}
