% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HorF.R
\name{plot_HorF}
\alias{plot_HorF}
\title{Create the plots about Hamd and Fst.}
\usage{
plot_HorF(x, y, breaks = default_break, label = default_label)
}
\arguments{
\item{x}{a list including the allele frequencies of three populations(like: EUROPE, AMERICA, AFRICA).}

\item{y}{can be either "Fst" or "Hamd"}

\item{breaks}{vectors to categorize Fst or Hamd.}

\item{label}{customize the label of the legend.}
}
\value{
The list should include allele frequencies of three populations.
}
\description{
\code{plot_HorF} returns plots about Fst and Hadm at different locus.
}
\details{
Users can use this function to create the plot about either Fst or Hamd, and also can customize the breaks, which is
  a argument to categorize Fst and Hamd, as well as the label of the legend.
}
\examples{
plot_HorF("D2S1399", y="Hadm")
plot_HorF("D2S1399", y="Hadm", breaks = c(0, 0.8371, 0.8504, 0.8578, 0.8620, 0.8661, 0.8702, 0.8741,
 0.8781, 1), label = c(0, 0.8371, 0.8504, 0.8578, 0.8620, 0.8661, 0.8702, 0.8741, 0.8781))
plot_HorF("GATA101G01_12", y="Fst")
plot_HorF("GATA101G01_12", y="Fst", breaks = c(-1, 0.00018, 0.00086, 0.00242, 0.00617, 0.01318, 0.02512, 
0.04713, 0.13772, 1), label = c(0, 0.00018, 0.00086, 0.00242, 0.00617, 0.01318, 0.02512, 0.04713, 0.13772))
}
