% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aashind.R
\name{ptf_aashind}
\alias{ptf_aashind}
\title{AASHTO Group Index}
\usage{
ptf_aashind(.sieveno200, .ll, .pi, .aashtocl = NULL)
}
\arguments{
\item{.sieveno200}{A numeric vector containing \% Passing #200 Sieve.}

\item{.ll}{A numeric vector containing Liquid Limit.}

\item{.pi}{A numeric vector containing Plasticity Index.}

\item{.aashtocl}{optional: A character vector containing AASHTO Class 
(to identify A-2-6, A-2-7 & A-8). default: NULL;}
}
\value{
A numeric vector containing calculated AASHTO Group Index.
}
\description{
R pedotransfer function derived from AASHTO Group Index NASIS calculation.
}
\examples{
ptf_aashind(.sieveno200 = 60, .ll = 44, .pi = 26, .aashtocl = "A-7-6")
}
\author{
Andrew G. Brown.
}
