% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formative-elements.R
\name{SubGroupFormativeElements}
\alias{SubGroupFormativeElements}
\title{Extract Subgroup Formative Elements}
\usage{
SubGroupFormativeElements(x)
}
\arguments{
\item{x}{vector of subgroup taxa}
}
\value{
a \code{list}
}
\description{
Extract Subgroup Formative Elements
}
\author{
D.E. Beaudette
}
