% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formative-elements.R
\name{SubOrderFormativeElements}
\alias{SubOrderFormativeElements}
\title{Extract Suborder Formative Elements}
\usage{
SubOrderFormativeElements(x)
}
\arguments{
\item{x}{vector of suborder taxa}
}
\value{
a \code{list}
}
\description{
Extract Suborder Formative Elements
}
\author{
D.E. Beaudette
}
