% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formative-elements.R
\name{GreatGroupFormativeElements}
\alias{GreatGroupFormativeElements}
\title{Extract Greatgroup Formative Elements}
\usage{
GreatGroupFormativeElements(x)
}
\arguments{
\item{x}{vector of greatgroup taxa}
}
\value{
a \code{list}
}
\description{
Extract Greatgroup Formative Elements
}
\author{
D.E. Beaudette
}
