% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_classes.R
\docType{package}
\name{lassoinf-package}
\alias{lassoinf}
\alias{lassoinf-package}
\title{lassoinf: Selective Inference for the Randomized Lasso}
\description{
Post-selection confidence intervals and p-values for the
randomized lasso. Start with \code{\link{LassoInference}}; see
\code{vignette("gaussian_lasso", package = "lassoinf")}.
}
\references{
Lee, J. D., Sun, D. L., Sun, Y. and Taylor, J. E. (2016). Exact
post-selection inference, with application to the lasso.
\emph{The Annals of Statistics}, 44(3), 907--927.
\doi{10.1214/15-AOS1371}

Tian, X. and Taylor, J. (2018). Selective inference with a randomized
response. \emph{The Annals of Statistics}, 46(2), 679--710.
\doi{10.1214/17-AOS1564}

Panigrahi, S., Fry, K. and Taylor, J. (2024). Exact selective inference
with randomization. \emph{Biometrika}, 111(4), 1109--1127.
\doi{10.1093/biomet/asae019}

Liu, S. (2023). An exact sampler for inference after polyhedral model
selection. \emph{arXiv preprint}. \doi{10.48550/arXiv.2308.10346}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonathan-taylor/lassoinf}
  \item Report bugs at \url{https://github.com/jonathan-taylor/lassoinf/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Taylor \email{jtaylo@stanford.edu} [copyright holder]

Authors:
\itemize{
  \item Jonathan Taylor \email{jtaylo@stanford.edu} [copyright holder]
  \item Kevin Fry
  \item Snigdha Panigrahi
  \item Alden Green
  \item Sifan Liu (truncated bivariate normal computations, see inst/COPYRIGHTS) [copyright holder]
}

}
\keyword{internal}
