% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{map}
\alias{map}
\title{Map an R function across SciDB streaming data frame chunks.}
\usage{
map(f, final, convertFactor = as.integer)
}
\arguments{
\item{f}{a function of a single data frame input argument that returns a data frame
output. The output data frame column types must match the SciDB stream operator
'types' argument.}

\item{final}{optional function applied to last output value before returning. If supplied, \code{final} must be a function of a
single data frame that returns a data frame compatible with the expected types (just like \code{f}).}

\item{convertFactor}{a function for conversion of R factor values into one of double, integer, or character for return to SciDB.}
}
\description{
The SciDB streaming API works with R functions that take a data frame input value
and produce a data frame output value. The output data frame column types must match the
types declared in the SciDB stream operator.
}
\note{
Factor and logical values are converted by default into integer values. Set
\code{convertFactor=as.character} to convert factor values to character strings instead.

Nothing is returned to SciDB when then function \code{f} returns \code{NULL}. Use this in combination
with the \code{final} function to perform aggregation across chunks (see the examples).
}
\examples{
# (Run all the examples from a command line)
# Identity function:
# iquery -aq "stream(build(<val:double> [i=1:5,5,0], i), 'R --slave -e \\"library(scidbstrm); map(I)\\"', 'format=df', 'types=double')"

# Three supported types so far are double, string, int32.
# NAs work for all of them and are translated to SciDB NULL.
# iquery -aq "stream(apply(build(<x:double> [i=1:5,5,0], i), y, 'cazart', z, int32(x)), 'R --slave -e \\"library(scidbstrm); map(function(x) {x[3,] = NA;x})\\"', 'format=df', 'types=double,string,int32')"

# See more examples in the following directory:
system.file('examples', package='scidbstrm')
}
\seealso{
\code{\link{schema}}
}

