% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported.R
\name{run}
\alias{run}
\title{Process SciDB streaming data frame chunks}
\usage{
run(f, convert_factor = as.integer)
}
\arguments{
\item{f}{a function of a single data frame input argument that returns a data frame
output. The output data frame column types must match the SciDB stream operator
'types' argument.}

\item{convert_factor}{a function for conversion of R factor values into a supported type: one of double, integer, or character.}
}
\description{
The SciDB streaming API works with R functions that map a data frame input value
to a data frame output value. The output data frame column types must match the
}
\note{
Factor and logical values are converted by default into integer values. Set
\code{convert_factor=as.character} to convert factor values to character strings instead.
}
\examples{
\dontrun{
#  Run from a command-line
# iquery -aq "stream(build(<val:double> [i=1:5,5,0], i), 'R --slave -e \\"library(scidbstrm); run(I)\\"', 'format=df', 'types=double')"
}
}
\seealso{
\code{\link{schema}}
}

