% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formative-elements.R
\name{SubGroupFormativeElements}
\alias{SubGroupFormativeElements}
\title{Identify subgroup formative elements within subgroup or lower taxa.}
\usage{
SubGroupFormativeElements(x)
}
\arguments{
\item{x}{subgroup or lower taxa}
}
\description{
Identify subgroup formative elements within subgroup or lower taxa.
}
